hello_gcp1/requirements.txt
requests==2.24.0
google-cloud-pubsub